refactor: move ui elements to `ui` package. add material3 expressive LoadingIndicator
